+++ src/session/src/Handler/RedisHandler.php
@@ -1,15 +1,13 @@
 <?php
 
-namespace Mix\Session;
+namespace Mix\Session\Handler;
 
 use Mix\Redis\Redis;
-use Mix\Session\Handler\HandlerInterface;
 
 /**
  * Class RedisHandler
- * @package Mix\Session
+ * @package Mix\Session\Handler
  * @author liu,jian <coder.keda@gmail.com>
- * @deprecated 废弃，请使用 Handler 目录内的 Handler Class
  */
 class RedisHandler implements HandlerInterface
 {
